+++ src/components/component.include.ts
@@ -1,0 +1,21 @@
+/**
+ * component.include.ts: Component inclusdion engine.
+ * Uses: component.include.json
+ * This script loads templates used in project, at `template` and `templateString`, and generates `components.ts` file (to be imported by project).
+ * This script is used at build time (in `webpack.config.ts`).
+ */
+
+import fs from 'node:fs';
+import path from 'node:path';
+import components from './component.include.json';
+console.log('Loading Web components ...');
+const templates =
+    '\n' +
+    components.include.map((component) => fs.readFileSync(path.join(__dirname, component, component + '.template.html'), 'utf-8')).join('\n') +
+    '\n';
+fs.writeFileSync(
+    path.join(__dirname, 'components.ts'),
+    `/* Web components. File generated at ${new Date().toUTCString()}. */\n${components.include.map((component) => `export * from './${component}/${component}.component';`).join('\n')}`,
+);
+export { components, templates };
+console.log('Loading Web components done.');