+++ src/components/components.ts
@@ -1,0 +1,3 @@
+/* Web components. File generated at Wed, 05 Feb 2025 17:26:43 GMT. */
+export * from './collapsible-content-panel/collapsible-content-panel.component';
+export * from './fixed-scroll-panel/fixed-scroll-panel.component';