+++ src/createHTMLElement.ts
@@ -1,8 +1,8 @@
 export default function createHTMLElement<K extends keyof HTMLElementTagNameMap>(
     tag: K,
     parent?: HTMLElement | null,
-    i?: string | null,
-    c?: string | null,
+    id?: string | null,
+    className?: string | null,
 ): HTMLElementTagNameMap[K] {
     if (arguments.length == 1) {
         return document.createElement(tag);
@@ -11,8 +11,8 @@ export default function createHTMLElement<K extends keyof HTMLElementTagNameMap>
         if (parent) {
             parent.append(result);
         }
-        if (arguments.length > 2 && i) result.setAttribute('id', i);
-        if (arguments.length > 3 && c) result.setAttribute('class', c);
+        if (arguments.length > 2 && id) result.setAttribute('id', id);
+        if (arguments.length > 3 && className) result.setAttribute('class', className);
         return result;
     }
 }