+++ src/main.ts
@@ -1,22 +1,20 @@
-// declare module "*.scss"
-
 import evalInput from './evalInput';
 import evalPrompt from './evalPrompt';
 import { Shell } from './Shell';
+import './components/components';
 import './main.css';
 
-// import './components/simple-nav/simple-nav.component';
-import './components/collapsible-content-panel/collapsible-content-panel.component';
+declare global {
+    var ShellPointer: Shell;
+}
 
-export const input: string = '';
-
-function bootstrap() {
-    Shell.initialize({
+async function bootstrap(): Promise<void> {
+    global.ShellPointer = await Shell.initialize({
         containerId: 'mathjslab-prompt',
         examplesId: 'mathjslab-examples',
         evalPrompt,
         evalInput,
-        input,
+        input: '',
     });
 }
 bootstrap();